+++ packages/cubejs-backend-shared/src/env.ts
@@ -1481,6 +1481,8 @@ const variables: Record<string, (...args: any) => any> = {
     .asString(),
   cacheAndQueueDriver: () => get('CUBEJS_CACHE_AND_QUEUE_DRIVER')
     .asString(),
+  defaultApiScope: () => get('CUBEJS_DEFAULT_API_SCOPES')
+    .asArray(','),
   jwkUrl: () => get('CUBEJS_JWK_URL')
     .asString(),
   jwtKey: () => get('CUBEJS_JWT_KEY')